fix: simultaneous playing sound in wikipedia
click the audio link repeatly in wikipedia will cuase the sound to play simultaneously

[autofix.ci] apply automated fixes
