+++ src/ffmpegaudio.hh
@@ -2,18 +2,34 @@
 #define __FFMPEGAUDIO_HH_INCLUDED__
 
 #ifdef MAKE_FFMPEG_PLAYER
-
+  #include "audiooutput.hh"
   #include <QObject>
   #include <QMutex>
-  #include <QAtomicInt>
   #include <QByteArray>
   #include <QThread>
+extern "C" {
+  #include <libavcodec/avcodec.h>
+  #include <libavformat/avformat.h>
+  #include <libavutil/avutil.h>
+  #include "libswresample/swresample.h"
+}
 
+  #include <QString>
+  #include <QDataStream>
+
+  #include <vector>
+  #if ( QT_VERSION >= QT_VERSION_CHECK( 6, 2, 0 ) )
+    #include <QMediaDevices>
+
+    #include <QAudioDevice>
+  #endif
+using std::vector;
 namespace Ffmpeg {
-
+class DecoderThread;
 class AudioService: public QObject
 {
   Q_OBJECT
+  std::shared_ptr< DecoderThread > thread;
 
 public:
   static AudioService & instance();
@@ -29,6 +45,43 @@ private:
   ~AudioService();
 };
 
+struct DecoderContext
+{
+  enum {
+    kBufferSize = 32768
+  };
+
+  static QMutex deviceMutex_;
+  QAtomicInt & isCancelled_;
+  QByteArray audioData_;
+  QDataStream audioDataStream_;
+  AVFormatContext * formatContext_;
+  #if LIBAVCODEC_VERSION_MAJOR < 59
+  AVCodec * codec_;
+  #else
+  const AVCodec * codec_;
+  #endif
+  AVCodecContext * codecContext_;
+  AVIOContext * avioContext_;
+  AVStream * audioStream_;
+  AudioOutput * audioOutput;
+  bool avformatOpened_;
+
+  SwrContext * swr_;
+
+  DecoderContext( QByteArray const & audioData, QAtomicInt & isCancelled );
+  ~DecoderContext();
+
+  bool openCodec( QString & errorString );
+  void closeCodec();
+  bool openOutputDevice( QString & errorString );
+  void closeOutputDevice();
+  bool play( QString & errorString );
+  void stop();
+  bool normalizeAudio( AVFrame * frame, vector< uint8_t > & samples );
+  void playFrame( AVFrame * frame );
+};
+
 class DecoderThread: public QThread
 {
   Q_OBJECT
@@ -36,6 +89,7 @@ class DecoderThread: public QThread
   static QMutex deviceMutex_;
   QAtomicInt isCancelled_;
   QByteArray audioData_;
+  DecoderContext d;
 
 public:
   DecoderThread( QByteArray const & audioData, QObject * parent );
